# Use Singular maMapPoly to avoid quadratic complexity in polynomial evaluation