### :books: Description

Currently evaluation of mutivariate polynomials using the Singular implementation is dominated by a quadratic algorithm for large polynomials, making it less efficient than the generic implementation.

Another Singular API does not have this problem and performs a more classical computation.

Fixes issue #35374 

Example:
```
sage: p = next_prime(2**24)
sage: x, y = polygens(GF(p), list("xy"))
sage: Rgeneric = PolynomialRing(GF(p), names=["x","y"], implementation="generic")
sage: pol = (x+y+1)^300
sage: polg = Rgeneric(pol)
# Generic
sage: %timeit polg(12,34)
86.5 ms ± 828 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# Before patch
sage: %time pol(12,34)
CPU times: user 6.76 s, sys: 30 ms, total: 6.79 s
# After patch
sage: %timeit pol(12,34)
12 ms ± 635 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

### :memo: Checklist

- [x] The title is concise, informative, and self-explanatory.
- [x] The description explains in detail what this PR is about.
- [x] I have linked a relevant issue or discussion.
- [ ] I have created tests covering the changes.
- [ ] I have updated the documentation accordingly.